belim ada validasi expired

+++ php/deleteData.php
@@ -2,9 +2,12 @@
 
 header('Content-Type: application/json');
 include "./config.php";
+session_start();
+
+$userId = $_SESSION["id_user"];
 
 $id = $_POST['id'];
-$sql = "DELETE FROM events_dev WHERE id = '$id'";
+$sql = "DELETE FROM events_dev WHERE id = '$id' AND id_user = '$userId'";
 $results = $db->query($sql);
 
 $message = "Kegiatan berhasil dihapus!";
